refactor project structure and introduce 'util' package
