+++ app/src/main/java/com/example/palliativecareapplication/ui/MainActivity.kt
@@ -1,9 +1,10 @@
-package com.example.palliativecareapplication
+package com.example.palliativecareapplication.ui
 
 import androidx.appcompat.app.AppCompatActivity
 import android.os.Bundle
 import androidx.fragment.app.Fragment
 import androidx.fragment.app.FragmentActivity
+import com.example.palliativecareapplication.R
 
 class MainActivity : AppCompatActivity() {
 
@@ -30,6 +31,6 @@ class MainActivity : AppCompatActivity() {
     override fun onCreate(savedInstanceState: Bundle?) {
         super.onCreate(savedInstanceState)
         setContentView(R.layout.activity_main)
-        swipeFragment(this,ViewTopicsFragment())
+        swipeFragment(this, ViewTopicsFragment())
     }
 }
