+++ app/src/main/java/com/example/palliativecareapplication/ui/adapter/TopicAdapter.kt
@@ -1,4 +1,4 @@
-package com.example.palliativecareapplication.adapter
+package com.example.palliativecareapplication.ui.adapter
 
 import android.app.AlertDialog
 import android.content.Context
@@ -8,52 +8,43 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.FragmentActivity
-import androidx.recyclerview.widget.GridLayoutManager
-import androidx.recyclerview.widget.LinearLayoutManager
 import androidx.recyclerview.widget.RecyclerView
-import com.example.palliativecareapplication.MainActivity
-import com.example.palliativecareapplication.databinding.CardPostBinding
-import com.example.palliativecareapplication.databinding.FragmentViewPostsBinding
-import com.example.palliativecareapplication.model.Attachment
+import com.example.palliativecareapplication.ui.MainActivity
+import com.example.palliativecareapplication.ui.TopicDetailsFragment
+import com.example.palliativecareapplication.databinding.CardTopicBinding
 import com.example.palliativecareapplication.model.FirebaseNames
-import com.example.palliativecareapplication.model.Post
 import com.example.palliativecareapplication.model.Topic
 import com.google.firebase.firestore.FirebaseFirestore
 import com.google.firebase.firestore.ktx.firestore
 import com.google.firebase.ktx.Firebase
 import com.squareup.picasso.Picasso
 
-class PostAdapter(var data: ArrayList<Post>, var topic: Topic, var parentBinding: FragmentViewPostsBinding) :
-    RecyclerView.Adapter<PostAdapter.MyViewHolder>() {
+class TopicAdapter(var data: ArrayList<Topic>): RecyclerView.Adapter<TopicAdapter.MyViewHolder>() {
     lateinit var context: Context
     lateinit var db: FirebaseFirestore
     private var initialData = data
 
 
-    class MyViewHolder(val cardViewBinding: CardPostBinding) :
-        RecyclerView.ViewHolder(cardViewBinding.root)
+    class MyViewHolder(val cardViewBinding: CardTopicBinding): RecyclerView.ViewHolder(cardViewBinding.root)
 
     override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): MyViewHolder {
         context = parent.context
         db = Firebase.firestore
 
-        val binding: CardPostBinding =
-            CardPostBinding.inflate(LayoutInflater.from(context), parent, false)
+        val binding : CardTopicBinding
+                = CardTopicBinding.inflate(LayoutInflater.from(context),parent,false)
         return MyViewHolder(binding)
     }
 
     override fun onBindViewHolder(holder: MyViewHolder, position: Int) {
         holder.cardViewBinding.apply {
             tvTitle.text = data[position].title
-            tvDetails.text = data[position].details
-            if (data[position].attachments?.isNotEmpty() == true) {
-                val attachmentsAdapter = AttachmentAdapter(data[position].attachments!!,topic, parentBinding)
-                rvAttachments.layoutManager = GridLayoutManager(context, 2)
-                rvAttachments.adapter = attachmentsAdapter
-            }
+            tvDoctorName.text = data[position].doctorName
+            tvFollowersCount.text = "${data[position].usersFollowing} متابعين"
+            Picasso.get().load(data[position].image).into(imgTopic)
 
             if(MainActivity.isPatient){
-                btnDelete.visibility = View.GONE
+                tvFollowersCount.visibility = View.GONE
             }
         }
 
@@ -62,6 +53,12 @@ class PostAdapter(var data: ArrayList<Post>, var topic: Topic, var parentBinding
         }
 
 
+        holder.cardViewBinding.root.setOnClickListener {
+            MainActivity.swipeFragment(context as FragmentActivity,
+                TopicDetailsFragment(data[position])
+            )
+        }
+
     }
 
     override fun getItemCount(): Int {
@@ -69,12 +66,13 @@ class PostAdapter(var data: ArrayList<Post>, var topic: Topic, var parentBinding
     }
 
 
+
     private fun showDeleteAlert(position: Int) {
         AlertDialog.Builder(context).apply {
-            setTitle("Delete Post")
-            setMessage("Are you sure that you want to delete this Post?")
+            setTitle("Delete topic")
+            setMessage("Are you sure that you want to delete this topic?")
             setPositiveButton("Yse") { _, _ ->
-                deletePostById(position)
+                deleteTopicById(position)
             }
             setCancelable(true)
             setNegativeButton("No") { dialogInterface: DialogInterface, _ ->
@@ -86,10 +84,8 @@ class PostAdapter(var data: ArrayList<Post>, var topic: Topic, var parentBinding
     }
 
 
-    private fun deletePostById(position: Int) {
-        db.collection(FirebaseNames.COLLECTION_TOPICS)
-            .document(topic.id).collection(FirebaseNames.COLLECTION_POSTS)
-            .document(data[position].id)
+    fun deleteTopicById(position: Int) {
+        db.collection(FirebaseNames.COLLECTION_TOPICS).document(data[position].id)
             .delete()
             .addOnSuccessListener { _ ->
                 Log.e("TAG", "Deleted Successfully")
@@ -102,11 +98,11 @@ class PostAdapter(var data: ArrayList<Post>, var topic: Topic, var parentBinding
     }
 
 
-    fun search(text: String) {
-        val newArray = initialData.filter { post ->
-            post.title.contains(text) // || post.doctorName.startsWith(text)
+    fun search(text : String){
+        val newArray = initialData.filter { topic ->
+            topic.title.contains(text) // || topic.doctorName.startsWith(text)
         }
-        data = newArray as ArrayList<Post>
+        data = newArray as ArrayList<Topic>
         notifyDataSetChanged()
     }
