+++ app/src/main/res/layout/fragment_add_topic.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/color_background"
-    tools:context=".AddTopicFragment">
+    tools:context=".ui.AddTopicFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/top_app_bar"
