+++ app/src/main/res/layout/fragment_view_posts.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:background="@color/color_background"
-    tools:context=".ViewPostsFragment">
+    tools:context=".ui.ViewPostsFragment">
 
     <com.google.android.material.appbar.AppBarLayout
         android:id="@+id/top_app_bar"
