+++ app/src/main/res/layout/fragment_view_topics.xml
@@ -5,7 +5,7 @@
     android:layout_height="match_parent"
     android:padding="16dp"
     android:background="@color/color_background"
-    tools:context="com.example.palliativecareapplication.ViewTopicsFragment">
+    tools:context="com.example.palliativecareapplication.ui.ViewTopicsFragment">
 
     <com.google.android.material.textfield.TextInputLayout
         android:id="@+id/container_search"
